Just added the toolbar panel
Because there wasn't any option to neither clear
the whole grid nor a button to simulate an eraser

+++ script.js
@@ -2,7 +2,9 @@
 let gridDimension = 16;
 let container = document.getElementById('container');
 let button = document.getElementById('pop-up');
-
+let clear = document.getElementById('clear');
+let eraser = document.getElementById('eraser');
+let paintingColor = 'black';
 
 function createGrid(dimension) {
     for (let i = 0; i < dimension; i++) {
@@ -23,28 +25,50 @@ function deletePreviousGrid() {
     }
 };
 
-function setGridDimension() {
-    gridDimension = prompt("Please enter the number of squares");
-    deletePreviousGrid();
-    createGrid(gridDimension);
-    paintGrid();
+function clearGrid(){
+    container.querySelectorAll('.item').forEach(item => {
+        item.style.backgroundColor = 'white';
+    });
 };
+
+function setGridDimension() {
+    let answer;
+    while (true) {
+        answer = prompt("Please enter the number of squares");
+        if (answer <= 100 && answer >= 1) {
+            deletePreviousGrid();
+            createGrid(answer);
+            paintGrid();
+            break;
+        } else {
+            alert("Please enter a valid number");
+        }
+    }
+};
+
 
 function paintGrid() {
     container.querySelectorAll('.item').forEach(item => {
 
         item.addEventListener('mouseenter', (event) => {
             if (event.buttons === 1) {
-                event.target.style.backgroundColor = 'black';
+                event.target.style.backgroundColor = paintingColor;
             }
         });
 
         item.addEventListener('mousedown', () => {
-            item.style.backgroundColor = 'black';
+            item.style.backgroundColor = paintingColor;
         });
     });
-
 }
 
 createGrid(16);
 button.addEventListener('click', setGridDimension);
+
+eraser.addEventListener('click', () =>{
+    paintingColor = 'white';
+});
+
+clear.addEventListener('click', ()=>{
+    clearGrid();
+});
